Move dump-parse-tree.h from semantics into parser.
