+++ tools/f18/CMakeLists.txt
@@ -36,5 +36,4 @@ add_executable(f18-parse-demo
 
 target_link_libraries(f18-parse-demo
   FortranParser
-  FortranSemantics
 )
